requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
redis==5.0.1
hiredis==2.3.2
celery==5.3.6
flower==2.0.1
websockets==12.0
uuid7==0.1.0
httpx==0.26.0
aioboto3==12.3.0
firebase-admin==6.5.0
aio-pika==9.3.1
slowapi==0.1.9